requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.3
django-allauth==0.50.0
django-paypal==2.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
postgres==4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
